fix: restructure tests
